perf: split runtime deps and dev deps
