+++ deps.ts
@@ -1,12 +1,5 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 export {
-  assert,
-  assertEquals,
-  assertNotEquals,
-  assertRejects,
-  assertThrows,
-} from "https://deno.land/std@0.191.0/testing/asserts.ts";
-export {
   isRedirectStatus,
   type RedirectStatus,
   Status,
@@ -23,6 +16,4 @@ export {
   type OAuth2ClientConfig,
   type Tokens,
 } from "https://deno.land/x/oauth2_client@v1.0.0/mod.ts";
-export { walk } from "https://deno.land/std@0.191.0/fs/walk.ts";
-export { globToRegExp } from "https://deno.land/std@0.191.0/path/glob.ts";
 export { SECOND } from "https://deno.land/std@0.191.0/datetime/constants.ts";
