+++ dev_deps.ts
@@ -1,0 +1,11 @@
+// Copyright 2023 the Deno authors. All rights reserved. MIT license.
+export {
+  assert,
+  assertEquals,
+  assertNotEquals,
+  assertRejects,
+  assertThrows,
+} from "https://deno.land/std@0.191.0/testing/asserts.ts";
+export { walk } from "https://deno.land/std@0.191.0/fs/walk.ts";
+export { globToRegExp } from "https://deno.land/std@0.191.0/path/glob.ts";
+export * from "./deps.ts";
