+++ src/get_session_id_test.ts
@@ -1,5 +1,5 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
-import { assertEquals } from "../deps.ts";
+import { assertEquals } from "../dev_deps.ts";
 import { SITE_COOKIE_NAME } from "./_core.ts";
 import { getSessionId } from "./get_session_id.ts";
