+++ src/sign_out_test.ts
@@ -1,5 +1,5 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
-import { assertEquals, Status, type Tokens } from "../deps.ts";
+import { assertEquals, Status, type Tokens } from "../dev_deps.ts";
 import { signOut } from "./sign_out.ts";
 import {
   deleteStoredTokensBySiteSession,
